# tests/fixtures/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "2009scape Wiki API",
...
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/types": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List every entity type and how its fields are presented",
        "description": "List every type with its attributes and its relationships, labelled for reading\neach way round.",
        "operationId": "types",
        "responses": {
          "200": {
            "description": "Every type served, with its attributes and relationships.",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TypeInfo"
                  },
                  "type": "array",
                  "title": "Response Types"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/types/{entity_type}/entities": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "List all entities of one type, a page at a time",
        "description": "List every published entity of one type, sorted and paged, leaving out\nvariants.",
        "operationId": "listing",
        "parameters": [
          {
            "name": "entity_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which type of thing to list, such as `item` or `quest`. `GET /v1/types` lists the types this build serves."
            },
            "description": "Which type of thing to list, such as `item` or `quest`. `GET /v1/types` lists the types this build serves."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume.",
              "default": 50,
              "title": "Limit"
            },
            "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "How many results to skip before this page. To read the next page, pass back the `next_offset` from the last response. It is null once you have reached the end.",
              "default": 0,
              "title": "Offset"
            },
            "description": "How many results to skip before this page. To read the next page, pass back the `next_offset` from the last response. It is null once you have reached the end."
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "description": "How to sort the listing: `name` for alphabetical, `id` for the game's own numbering.",
              "default": "name"
            },
            "description": "How to sort the listing: `name` for alphabetical, `id` for the game's own numbering."
          }
        ],
        "responses": {
          "200": {
            "description": "One page of the index for that type.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntitySummaryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/search": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Search the full text across every type",
        "description": "Search names, alternative names and descriptions across every type, best first\nwith a score.",
        "operationId": "search",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "What to search for, typed the way a reader would type it. Every word is matched as a prefix and all of them have to match, so `drag scim` finds the dragon scimitar. There is no query syntax: quotes, brackets and words like `OR` count as ordinary text.",
              "title": "Q"
            },
            "description": "What to search for, typed the way a reader would type it. Every word is matched as a prefix and all of them have to match, so `drag scim` finds the dragon scimitar. There is no query syntax: quotes, brackets and words like `OR` count as ordinary text."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit the answer to certain types of thing. Repeat the parameter for more than one, as in `?type=item&type=npc`. Leave it off to cover every type.",
              "title": "Type"
            },
            "description": "Limit the answer to certain types of thing. Repeat the parameter for more than one, as in `?type=item&type=npc`. Leave it off to cover every type."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume.",
              "default": 50,
              "title": "Limit"
            },
            "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume."
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "How many results to skip before this page. To read the next page, pass back the `next_offset` from the last response. It is null once you have reached the end.",
              "default": 0,
              "title": "Offset"
            },
            "description": "How many results to skip before this page. To read the next page, pass back the `next_offset` from the last response. It is null once you have reached the end."
          }
        ],
        "responses": {
          "200": {
            "description": "One page of matches, best first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/find": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Look up the one entity that goes by this name",
        "description": "Turn a name into the one entity it means, with the ranked runners-up.\n\n`best_match` is null only when nothing matched at all.",
        "operationId": "find",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The full name of the single thing you mean, such as `Dragon scimitar`.",
              "title": "Name"
            },
            "description": "The full name of the single thing you mean, such as `Dragon scimitar`."
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityType"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Limit the answer to certain types of thing. Repeat the parameter for more than one, as in `?type=item&type=npc`. Leave it off to cover every type.",
              "title": "Type"
            },
            "description": "Limit the answer to certain types of thing. Repeat the parameter for more than one, as in `?type=item&type=npc`. Leave it off to cover every type."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume.",
              "default": 50,
              "title": "Limit"
            },
            "description": "How many results one page holds. The response repeats the limit it used, so you never have to assume."
          }
        ],
        "responses": {
          "200": {
            "description": "The single best match, if there is one, plus the runners-up.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Match"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    },
    "/v1/near-names": {
      "get": {
        "tags": [
          "discovery"
        ],
        "summary": "Ask which real names a misspelt one might have meant",
        "description": "Turn an unmatched name into the real names it may have meant, each carrying\nidentity and nothing else.\n\nAn empty answer means nothing was close, which is an answer rather than a failure.",
        "operationId": "near_names",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The name that answered to nothing, exactly as it was typed, such as `dragon scimtar`.",
              "title": "Name"
            },
            "description": "The name that answered to nothing, exactly as it was typed, such as `dragon scimtar`."
          },
          {
            "name": "type",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EntityType",
              "description": "Which sort of thing the name belongs to. Required: `dagon scimitar` and `dagon` are near misses for different things depending on whether an item or an npc was meant, so this question cannot be answered without it. `GET /v1/types` lists the sorts this build serves."
            },
            "description": "Which sort of thing the name belongs to. Required: `dagon scimitar` and `dagon` are near misses for different things depending on whether an item or an npc was meant, so this question cannot be answered without it. `GET /v1/types` lists the sorts this build serves."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 25,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "description": "At most how many close names to offer. Deliberately small: this is a question to put to a person, not a listing to read.",
              "title": "Limit"
            },
            "description": "At most how many close names to offer. Deliberately small: this is a question to put to a person, not a listing to read."
          },
          {
            "name": "keep",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number",
                  "maximum": 1.0,
                  "minimum": 0.0
                },
                {
                  "type": "null"
                }
              ],
              "description": "How close to the best candidate a name has to be to be offered alongside it, as a share of the best score. `1.0` offers only names that matched exactly as well; lower values widen the field.",
              "title": "Keep"
            },
            "description": "How close to the best candidate a name has to be to be offered alongside it, as a share of the best score. `1.0` offers only names that matched exactly as well; lower values widen the field."
          }
        ],
        "responses": {
          "200": {
            "description": "The closest names that exist, best first, or nothing at all when none of them is close enough to be worth offering.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResultPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "No key was sent, or it is not one this service answers."
          },
          "403": {
            "description": "Too many refused requests have come from this address."
          },
          "429": {
            "description": "This key has asked for more than its share; `Retry-After` says when to come back."
          }
        },
        "security": [
          {
            "issued_key": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AttributeSpec": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "group": {
            "$ref": "#/components/schemas/AttributeGroup"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "format": {
            "$ref": "#/components/schemas/AttributeFormat"
          },
          "unit": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Unit"
              },
              {
                "type": "null"
              }
            ]
          },
          "display": {
            "type": "boolean",
            "title": "Display",
            "default": true
          },
          "derived": {
            "type": "boolean",
            "title": "Derived",
            "default": false
          },
          "prominent": {
            "type": "boolean",
            "title": "Prominent",
            "default": false
          },
          "choices": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Choices"
          }
        },
        "type": "object",
        "required": [
          "key",
          "label",
          "group",
          "order",
          "format"
        ],
        "title": "AttributeSpec",
        "description": "One attribute as declared: label, group, order, format, unit, and any fixed set\nof values it takes."
      },
...
      "EntitySummary": {
        "properties": {
          "link": {
            "$ref": "#/components/schemas/Link"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "link",
          "type"
        ],
        "title": "EntitySummary",
        "description": "One entity as it appears in a list: where it points, and a line about it."
      },
      "EntityType": {
        "type": "string",
        "title": "EntityType",
        "description": "The types of thing this knowledge base holds: items, NPCs, shops and more. Published as an open string rather than a fixed enum, because new values get added as the data grows. The ones this build knows are listed under `x-known-values`. Treat anything else as a value you cannot render yet, not as an error.",
        "x-known-values": [
          "item",
          "npc",
          "shop",
          "quest",
          "location"
        ],
        "examples": [
          "item"
        ]
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Link": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "id": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "minLength": 1,
            "title": "Slug"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "icon_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Ref"
          }
        },
        "type": "object",
        "required": [
          "type",
          "id",
          "slug",
          "label"
        ],
        "title": "Link",
        "description": "A pointer to another entity, carrying `type`, `id`, `slug` and `label` but never\na URL."
      },
...
      "Match": {
        "properties": {
          "best_match": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Link"
              },
              {
                "type": "null"
              }
            ]
          },
          "results": {
            "$ref": "#/components/schemas/SearchResultPage"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "Match",
        "description": "The entity a name most likely meant, null when nothing matched, with the ranked\ncandidates either way."
      },
...
      "EntitySummaryPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/EntitySummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "readOnly": true
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more",
          "next_offset"
        ],
        "title": "Page[EntitySummary]"
      },
...
      "SearchResultPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "maximum": 200.0,
            "minimum": 1.0,
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Offset"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "readOnly": true
          },
          "next_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Offset",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "has_more",
          "next_offset"
        ],
        "title": "Page[SearchResult]"
      },
...
      "RelationshipSpec": {
        "properties": {
          "rel": {
            "$ref": "#/components/schemas/RelationshipType"
          },
          "forward_label": {
            "type": "string",
            "title": "Forward Label"
          },
          "inverse_label": {
            "type": "string",
            "title": "Inverse Label"
          },
          "src_types": {
            "items": {
              "$ref": "#/components/schemas/EntityType"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Src Types"
          },
          "dst_types": {
            "items": {
              "$ref": "#/components/schemas/EntityType"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Dst Types"
          },
          "edge_attributes": {
            "items": {
              "$ref": "#/components/schemas/AttributeSpec"
            },
            "type": "array",
            "title": "Edge Attributes"
          },
          "group": {
            "$ref": "#/components/schemas/RelationshipGroup"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          }
        },
        "type": "object",
        "required": [
          "rel",
          "forward_label",
          "inverse_label",
          "src_types",
          "dst_types",
          "edge_attributes",
          "group",
          "order"
        ],
        "title": "RelationshipSpec",
        "description": "One relationship as declared: a label each direction, the types it joins, and\nthe attributes a link may carry."
      },
...
      "SearchResult": {
        "properties": {
          "link": {
            "$ref": "#/components/schemas/Link"
          },
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "score": {
            "type": "number",
            "minimum": 0.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "link",
          "type",
          "score"
        ],
        "title": "SearchResult",
        "description": "One entity a search matched; a higher `score` matched better."
      },
...
      "SortOrder": {
        "type": "string",
        "enum": [
          "name",
          "id"
        ],
        "title": "SortOrder",
        "description": "How a listing is sorted: alphabetically by name, or by id."
      },
...
      "TypeInfo": {
        "properties": {
          "type": {
            "$ref": "#/components/schemas/EntityType"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "plural": {
            "type": "string",
            "title": "Plural"
          },
          "order": {
            "type": "integer",
            "title": "Order"
          },
          "attributes": {
            "items": {
              "$ref": "#/components/schemas/AttributeSpec"
            },
            "type": "array",
            "title": "Attributes"
          },
          "relationships": {
            "items": {
              "$ref": "#/components/schemas/RelationshipSpec"
            },
            "type": "array",
            "title": "Relationships"
          }
        },
        "type": "object",
        "required": [
          "type",
          "label",
          "plural",
          "order",
          "attributes",
          "relationships"
        ],
        "title": "TypeInfo",
        "description": "One entity type, with every attribute and relationship declared for it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "issued_key": {
        "type": "http",
        "scheme": "bearer",
        "description": "Send the key you were issued as `Authorization: Bearer <token>`. Keys are handed out by whoever runs this deployment and do not expire; one that leaks is withdrawn by its holder's key id. A browser cannot keep one secret, so a public front end should call this through its own backend."
      }
    }
  },
...
}